+++ src/Schuss.cpp
@@ -3,7 +3,7 @@
 
 #include "Schuss.hpp"
 
-Schuss::Schuss(std::string filename, float angle, std::complex<float> position, float speed, SchussType type)
+Schuss::Schuss(std::string filename, double angle, std::complex<double> position, double speed, SchussType type)
 :Object(filename, SCHUSS, position.real(), position.imag(),
 speed*cos((angle+90)/180*M_PI),speed*sin((angle-90)/180*M_PI)), Type(type), Weight(1), Age(0)
 {
